requirements.txt
pytest
pytest-html
pytest-xdist
pytest-cov
allure-pytest
requests
